refactor(cache): 优化 cache 逻辑
 - 将各部分的 cache lib 统一整理到 @/lib/cache 中
 - 将适合的公开读取缓存迁移到 Next.js 16 Cache Components
 - 考虑到风险问题, 仍有部分 dashboard 的 unstable_cache 为迁移成 Cache Components
