+++ .serena/.gitignore
@@ -1,0 +1,2 @@
+/cache
+/project.local.yml
