+++ app/(admin)/dashboard/layout.tsx
@@ -8,6 +8,7 @@ import { headers } from "next/headers";
 import { Geist, Geist_Mono } from "next/font/google";
 import "@/app/globals.css";
 import { redirect } from "next/navigation";
+import { Suspense } from "react";
 import { Toaster } from "sonner";
 
 const geistSans = Geist({
@@ -20,30 +21,58 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-export default async function DashboardRootLayout({
+function DashboardLayoutFallback() {
+    return (
+        <div className="flex min-h-screen items-center justify-center text-sm text-muted-foreground">
+            正在加载后台...
+        </div>
+    );
+}
+
+async function DashboardAuthenticatedLayout({
     children,
 }: Readonly<{
     children: React.ReactNode;
 }>) {
-    // 获取用户的 Session
     const session = await auth.api.getSession({
-        headers: await headers() // you need to pass the headers object.
-    })
-
-    // 若未登录 跳转到 /sign-in
-    // 若 role 并非 admin, 跳转回主页
+        headers: await headers(),
+    });
 
     if (!session) {
-        return redirect("/sign-in")
+        return redirect("/sign-in");
     }
 
     if (session.user.role !== "admin") {
-        return redirect("/")
+        return redirect("/");
     }
 
     return (
+        <SidebarProvider>
+            <DashboardSidebar />
 
-        
+            <SidebarInset>
+                <div className="fixed top-5 right-5 z-50">
+                    <ModeToggle />
+                </div>
+
+                <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
+                    <div className="min-h-[100vh] flex-1 rounded-xl md:min-h-min">
+                        <div className="p-6">{children}</div>
+                    </div>
+                </div>
+
+                <Footer />
+            </SidebarInset>
+        </SidebarProvider>
+    );
+}
+
+export default function DashboardRootLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
         <html lang="zh-cn" suppressHydrationWarning>
             <body className={`${geistSans.variable} ${geistMono.variable} font-sans antialiased`}>
                 <ThemeProvider
@@ -51,33 +80,12 @@ export default async function DashboardRootLayout({
                     defaultTheme="system"
                     enableSystem
                     disableTransitionOnChange>
-
-                    <SidebarProvider>
-                        <DashboardSidebar />
-
-                        {/* 主要内容区域 - 不会被 Sidebar 遮挡 */}
-                        <SidebarInset>
-                            <div className="fixed top-5 right-5 z-50">
-                                <ModeToggle />
-                            </div>
-
-                            {/* 页面内容 */}
-                            <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
-                                <div className="min-h-[100vh] flex-1 rounded-xl md:min-h-min">
-                                    <div className="p-6">
-                                        {children}
-                                    </div>
-                                </div>
-                            </div>
-
-                            {/* 页脚 */}
-                            <Footer />
-                        </SidebarInset>
-                    </SidebarProvider>
-
+                    <Suspense fallback={<DashboardLayoutFallback />}>
+                        <DashboardAuthenticatedLayout>{children}</DashboardAuthenticatedLayout>
+                    </Suspense>
                 </ThemeProvider>
                 <Toaster richColors position="top-right" />
             </body>
         </html>
-    )
+    );
 }
