+++ app/(site)/blogs/[slug]/page.tsx
@@ -1,7 +1,19 @@
-import { getPublishedPostBySlug, getPostMetadataBySlug } from "@/lib/blog-cache";
+import {
+    getPublishedPostBySlug,
+    getPostMetadataBySlug,
+    getPublishedPosts,
+} from "@/lib/cache";
 import { notFound } from "next/navigation";
 import type { Metadata } from "next";
 import { BlogTableOfContents } from "@/components/blog-toc";
+
+export async function generateStaticParams() {
+    const posts = await getPublishedPosts();
+
+    return posts.map((post) => ({
+        slug: post.slug,
+    }));
+}
 
 // ─── SEO 元数据 ──────────────────────────────────────────
 export async function generateMetadata(
