+++ app/(site)/forms/page.tsx
@@ -1,4 +1,4 @@
-import { getActiveForms } from "@/lib/form-cache";
+import { getActiveForms } from "@/lib/cache";
 import { Badge } from "@/components/ui/badge";
 import { FileText, Lock, Users } from "lucide-react";
 import Link from "next/link";
