+++ app/(site)/layout.tsx
@@ -7,6 +7,7 @@ import { ThemeProvider } from "@/components/theme-provider";
 import { ModeToggle } from "@/components/ui/theme-toggle";
 import { Navbar } from "@/components/ui/navbar";
 import { UserAvatar } from "@/components/user-avatar";
+import { Suspense } from "react";
 import { Toaster } from "sonner";
 
 const geistSans = Geist({
@@ -23,6 +24,10 @@ export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
+
+function NavbarFallback() {
+  return <div className="mx-auto h-11 w-full max-w-[420px] rounded-full border bg-background/80" />;
+}
 
 export default function RootLayout({
   children,
@@ -44,7 +49,9 @@ export default function RootLayout({
 
           {/* 导航栏 */}
           <div className="pt-4">
-            <Navbar />
+            <Suspense fallback={<NavbarFallback />}>
+              <Navbar />
+            </Suspense>
           </div>
 
           {/* 主题切换按钮 - 固定在坐上角 */}
