+++ lib/cache/tags.ts
@@ -1,0 +1,10 @@
+export const CACHE_TAGS = {
+    BLOG_POSTS: "blog-posts",
+    ADMIN_BLOG_POSTS: "admin-blog-posts",
+    FORMS: "forms",
+    ADMIN_FORMS: "admin-forms",
+    FORM_SUBMISSIONS: "form-submissions",
+    REVIEW_SUBMISSIONS: "review-submissions",
+    REVIEW_RULES: "review-rules",
+    ADMIN_USERS: "admin-users",
+} as const;
