+++ outputs/runtime/vibe-sessions/2026-04-04-cache-refactor/cleanup-receipt.json
@@ -1,0 +1,14 @@
+{
+  "run_id": "2026-04-04-cache-refactor",
+  "stage": "cleanup",
+  "status": "completed",
+  "residuals": [
+    "lib/blog-cache.ts、lib/form-cache.ts、lib/review-cache.ts、lib/user-cache.ts 仍保留为 shim 兼容入口，不再承载独立缓存实现",
+    ".serena/ 为本地工具生成目录，未纳入本次缓存重构范围"
+  ],
+  "handoff_notes": [
+    "如果后续确认仓库内已无旧路径依赖，可再删除 shim 文件",
+    "如不希望 .serena/ 出现在工作区，可由用户决定是否加入忽略规则或清理"
+  ],
+  "final_verdict": "三阶段缓存重构已在 refactor/cache 分支完成，并通过 lint 与生产构建验证"
+}
